# openapi.yaml
openapi: 3.0.3
info:
  title: Garden Controller API
  version: 1.1.0
servers:
  - url: /
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Relay:
      type: object
      properties:
        channel: { type: integer, minimum: 1, maximum: 6 }
        state: { type: string, enum: [on, off] }
      required: [channel, state]
    PublishedSchedule:
      type: object
      properties:
        channel: { type: integer, minimum: 1, maximum: 6 }
        zone: { type: string }
        startTime: { type: string }
        durationSeconds: { type: integer, minimum: 1 }
      required: [channel, zone, startTime, durationSeconds]
paths:
  /health:
    get:
      responses:
        '200': { description: OK }
  /api/relays:
    get:
      responses:
        '200': { description: Desired/reported relay states }
  /api/state:
    get:
      responses:
        '200': { description: Full controller state }
  /api/commands:
    post:
      responses:
        '201': { description: Command queued }
        '400': { description: Invalid payload }
  /api/queue/next:
    get:
      parameters:
        - name: wait
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 25
          description: Optional long-poll wait duration in seconds.
      responses:
        '200': { description: Delivered command }
        '204': { description: No pending command (immediate or long-poll timeout) }
  /api/microcontroller/commands/{id}/ack:
    post:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Command acknowledged }
        '404': { description: Command not found }
        '409': { description: Command not yet delivered }
  /api/microcontroller/state:
    post:
      responses:
        '200': { description: Canonical telemetry persisted }
        '400': { description: Invalid payload }
  /api/microcontroller/relays/state:
    post:
      responses:
        '200': { description: Relay state accepted and persisted }
        '400': { description: Invalid payload }
  /api/microcontroller/schedules:
    post:
      responses:
        '200': { description: Schedules accepted and persisted }
        '400': { description: Invalid payload }
  /api/schedules:
    post:
      responses:
        '201': { description: Schedules stored and queued }
        '400': { description: Invalid payload }
  /api/news:
    get:
      responses:
        '200': { description: News list }
    post:
      responses:
        '201': { description: News created }
  /gui:
    get:
      security:
        - basicAuth: []
      responses:
        '200': { description: Management GUI }
  /gui/relays/{channel}/on:
    post:
      security:
        - basicAuth: []
      responses:
        '303': { description: Relay ON command queued }
        '400': { description: Invalid relay channel }
  /gui/relays/{channel}/off:
    post:
      security:
        - basicAuth: []
      responses:
        '303': { description: Relay OFF command queued }
        '400': { description: Invalid relay channel }
  /gui/schedules:
    post:
      security:
        - basicAuth: []
      responses:
        '303': { description: Schedule update queued }
        '400': { description: Invalid schedule payload }
  /openapi.json:
    get:
      responses:
        '200': { description: OpenAPI document as JSON }